refactor: improve exceptions module with SRP and comprehensive tests
- Split exceptions.rs into 4 focused modules:
  - models.rs: Exception and ExceptionsFile data structures
  - storage.rs: File I/O operations and path management
  - checker.rs: Business logic for exception checking and cleanup
  - interactive.rs: User interaction and exception handling workflow
- Move 60-line interactive processing logic from main.rs to handle_interactive_exceptions()
- Add 7 comprehensive tests covering all exception functionality
- Simplify main.rs to ~10 lines, focusing on entry point responsibilities
- All 10 tests pass with no warnings
- Maintain backward compatibility while improving maintainability
